feat: Se realiza el refactor del api
