+++ helpers/buscar_nombre_dedicacion.go
@@ -6,18 +6,20 @@ import (
 	"github.com/udistrital/resoluciones_docentes_mid/models"
 )
 
-func BuscarNombreDedicacion(id_dedicacion int) (nombre_dedicacion string) {
+func BuscarNombreDedicacion(id_dedicacion int) (nombre_dedicacion string, outputError map[string]interface{}) {
 	var nom_dedicacion string
 	query := "?limit=-1&query=Id:" + strconv.Itoa(id_dedicacion)
 	var dedicaciones []models.Dedicacion
-	if err2 := GetJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/dedicacion"+query, &dedicaciones); err2 == nil {
+	if response, err2 := GetJsonTest(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/dedicacion"+query, &dedicaciones); err2 == nil && response == 200{
 		if dedicaciones != nil {
 			nom_dedicacion = dedicaciones[0].Descripcion
 		} else {
 			nom_dedicacion = ""
 		}
-
+	}else{
+		outputError = map[string]interface{}{"funcion": "/BuscarNombreDedicacion", "err": err2.Error(), "status": "404"}
+		return nom_dedicacion, outputError
 	}
 
-	return nom_dedicacion
+	return nom_dedicacion, nil
 }
