+++ helpers/buscar_nombre_proveedor.go
@@ -7,20 +7,22 @@ import (
 	"github.com/udistrital/resoluciones_docentes_mid/models"
 )
 
-func BuscarNombreProveedor(DocumentoIdentidad int) (nombre_prov string) {
+func BuscarNombreProveedor(DocumentoIdentidad int) (nombre_prov string, outputError map[string]interface{}) {
 
 	var nom_proveedor string
 	queryInformacionProveedor := "?query=NumDocumento:" + strconv.Itoa(DocumentoIdentidad)
 	var informacion_proveedor []models.InformacionProveedor
-	if err2 := GetJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAgora")+"/"+beego.AppConfig.String("NscrudAgora")+"/informacion_proveedor/"+queryInformacionProveedor, &informacion_proveedor); err2 == nil {
+	if response, err2 := GetJsonTest(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAgora")+"/"+beego.AppConfig.String("NscrudAgora")+"/informacion_proveedor/"+queryInformacionProveedor, &informacion_proveedor); err2 == nil && response == 200 {
 		if informacion_proveedor != nil {
 			nom_proveedor = informacion_proveedor[0].NomProveedor
 		} else {
 			nom_proveedor = ""
 		}
-
+	}else{
+		outputError = map[string]interface{}{"funcion": "/BuscarNombreProveedor", "err": err2.Error(), "status": "404"}
+		return nom_proveedor, outputError
 	}
 
-	return nom_proveedor
+	return nom_proveedor, nil
 
 }
