+++ helpers/buscar_tipo_documento.go
@@ -7,10 +7,10 @@ import (
 	"github.com/udistrital/resoluciones_docentes_mid/models"
 )
 
-func BuscarTipoDocumento(Cedula string) (nombre_tipo_doc string) {
+func BuscarTipoDocumento(Cedula string) (nombre_tipo_doc string, outputError map[string]interface{}) {
 	var tipo_documento string
 	var temp []models.InformacionPersonaNatural
-	if err2 := GetJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAgora")+"/"+beego.AppConfig.String("NscrudAgora")+"/informacion_persona_natural/?limit=-1&query=Id:"+Cedula, &temp); err2 == nil {
+	if response, err2 := GetJsonTest(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAgora")+"/"+beego.AppConfig.String("NscrudAgora")+"/informacion_persona_natural/?limit=-1&query=Id:"+Cedula, &temp); err2 == nil && response == 200{
 		if temp != nil {
 			tipo_documento = temp[0].TipoDocumento.ValorParametro
 		} else {
@@ -19,8 +19,10 @@ func BuscarTipoDocumento(Cedula string) (nombre_tipo_doc string) {
 	} else {
 		fmt.Println("error en json", err2)
 		tipo_documento = "N/A"
+		outputError = map[string]interface{}{"funcion": "/BuscarTipoDocumento", "err": err2.Error(), "status": "404"}
+		return tipo_documento, outputError
 	}
 
-	return tipo_documento
+	return tipo_documento, nil
 
 }
