+++ helpers/desvinculaciones.go
@@ -9,7 +9,7 @@ import (
 	"github.com/udistrital/resoluciones_docentes_mid/models"
 )
 
-func AnularModificaciones(v []models.VinculacionDocente) (respuesta_total string) {
+func AnularModificaciones(v []models.VinculacionDocente) (outputError map[string]interface{}) {
 	var modRes []models.ModificacionResolucion
 	var respuesta_vinculacion string
 	var vinculacion_cancelada []models.VinculacionDocente
@@ -18,30 +18,30 @@ func AnularModificaciones(v []models.VinculacionDocente) (respuesta_total string
 	var respuesta_modificacion_vinculacion []models.ModificacionVinculacion
 
 	if response, err := GetJsonTest(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/modificacion_resolucion/?query=resolucionNueva:"+strconv.Itoa(v[0].IdResolucion.Id), &modRes); err == nil && response == 200 {
-		respuesta_total = "OK"
 	} else {
 		beego.Error(err)
-		respuesta_total = "Error"
+		outputError = map[string]interface{}{"funcion": "/AnularModificaciones6", "err": err.Error(), "status": "404"}
+		return outputError
 	}
 	for _, pos := range v {
 		//Se trae información de tabla de traza modificacion_vinculacion, para saber cuál vinculación hay que poner en true y cuál eliminar
 		query := "?limit=-1&query=ModificacionResolucion.Id:" + strconv.Itoa(modRes[0].Id) + ",VinculacionDocenteRegistrada.Id:" + strconv.Itoa(pos.Id)
 		if response2, err2 := GetJsonTest(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/modificacion_vinculacion"+query, &respuesta_modificacion_vinculacion); err2 == nil && response2 == 200 {
 			fmt.Println("modificacion_vinculacion", respuesta_modificacion_vinculacion)
-			respuesta_total = "OK"
 		} else {
 			beego.Error(err2)
-			respuesta_total = "Error"
+			outputError = map[string]interface{}{"funcion": "/AnularModificaciones5", "err": err2.Error(), "status": "404"}
+			return outputError
 		}
 
 		//se trae informacion de vinculación que fue cancelada
 		query2 := "?limit=-1&query=Id:" + strconv.Itoa(respuesta_modificacion_vinculacion[0].VinculacionDocenteCancelada.Id)
 		if response3, err3 := GetJsonTest(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/vinculacion_docente"+query2, &vinculacion_cancelada); err3 == nil && response3 == 200 {
 			fmt.Println("vinculacion_cancelada", vinculacion_cancelada)
-			respuesta_total = "OK"
 		} else {
 			beego.Error(err3)
-			respuesta_total = "Error"
+			outputError = map[string]interface{}{"funcion": "/AnularModificaciones4", "err": err3.Error(), "status": "404"}
+			return outputError
 		}
 		//se cambia a true vinculación que fue cancelada
 		vinculacion_cancelada[0].Estado = true
@@ -49,33 +49,33 @@ func AnularModificaciones(v []models.VinculacionDocente) (respuesta_total string
 		//Se le cambia estado en bd a vinculación cancelada
 		if err4 := SendJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/vinculacion_docente/"+strconv.Itoa(vinculacion_cancelada[0].Id), "PUT", &respuesta_vinculacion, vinculacion_cancelada[0]); err4 == nil {
 			fmt.Println("respuesta_vinculacion", respuesta_vinculacion)
-			respuesta_total = "OK"
 		} else {
 			beego.Error(err4)
-			respuesta_total = "Error"
+			outputError = map[string]interface{}{"funcion": "/AnularModificaciones3", "err": err4.Error(), "status": "404"}
+			return outputError
 		}
 
 		//se elimina registro en modificacion_vinculacion
 		if err5 := SendJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/modificacion_vinculacion/"+strconv.Itoa(respuesta_modificacion_vinculacion[0].Id), "DELETE", &respuesta_delete, respuesta_modificacion_vinculacion[0]); err5 == nil {
-			respuesta_total = "OK"
 		} else {
 			beego.Error(err5)
-			respuesta_total = "Error"
+			outputError = map[string]interface{}{"funcion": "/AnularModificaciones2", "err": err5.Error(), "status": "404"}
+			return outputError
 		}
 
 		//Se elimina vinculacion nueva
 		if err6 := SendJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/vinculacion_docente/"+strconv.Itoa(pos.Id), "DELETE", &respuesta_delete_vin, pos); err6 == nil {
 			fmt.Println("respuesta_eliminar_vin_nueva", respuesta_delete_vin)
-			respuesta_total = "OK"
 		} else {
 			beego.Error(err6)
-			respuesta_total = "Error"
+			outputError = map[string]interface{}{"funcion": "/AnularModificaciones1", "err": err6.Error(), "status": "404"}
+			return outputError
 		}
 	}
-	return respuesta_total
+	return 
 }
 
-func AnularAdicionDocente(v models.Objeto_Desvinculacion) (respuesta_total string) {
+func AnularAdicionDocente(v models.Objeto_Desvinculacion) (outputError map[string]interface{}) {
 	var respuesta_vinculacion string
 	var vinculacion_cancelada []models.VinculacionDocente
 	var respuesta_delete_vin string
@@ -86,18 +86,18 @@ func AnularAdicionDocente(v models.Objeto_Desvinculacion) (respuesta_total strin
 	query := "?limit=-1&query=ModificacionResolucion.Id:" + strconv.Itoa(v.IdModificacionResolucion) + ",VinculacionDocenteRegistrada.Id:" + strconv.Itoa(v.DocentesDesvincular[0].Id)
 	if response, err := GetJsonTest(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/modificacion_vinculacion"+query, &respuesta_modificacion_vinculacion); err == nil && response == 200 {
 		fmt.Println("modificacion_vinculacion", respuesta_modificacion_vinculacion)
-		respuesta_total = "OK"
 	} else {
-		respuesta_total = "error"
+		outputError = map[string]interface{}{"funcion": "/AnularAdicionDocente5", "err": err.Error(), "status": "404"}
+		return outputError
 	}
 
 	//se trae informacion de vinculacion que fue cancelada
 	query2 := "?limit=-1&query=Id:" + strconv.Itoa(respuesta_modificacion_vinculacion[0].VinculacionDocenteCancelada.Id)
 	if response2, err2 := GetJsonTest(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/vinculacion_docente"+query2, &vinculacion_cancelada); err2 == nil && response2 == 200 {
 		fmt.Println("vinculacion_cancelada", vinculacion_cancelada)
-		respuesta_total = "OK"
 	} else {
-		respuesta_total = "error"
+		outputError = map[string]interface{}{"funcion": "/AnularAdicionDocente4", "err": err2.Error(), "status": "404"}
+		return outputError
 	}
 	//se cambia a true vinculacion que fue cancelada
 	vinculacion_cancelada[0].Estado = true
@@ -107,35 +107,34 @@ func AnularAdicionDocente(v models.Objeto_Desvinculacion) (respuesta_total strin
 
 	if err3 := SendJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/vinculacion_docente/"+strconv.Itoa(vinculacion_cancelada[0].Id), "PUT", &respuesta_vinculacion, vinculacion_cancelada[0]); err3 == nil {
 		fmt.Println("respuesta_vinculacion", respuesta_vinculacion)
-		respuesta_total = "OK"
 	} else {
-		respuesta_total = "error"
+		outputError = map[string]interface{}{"funcion": "/AnularAdicionDocente3", "err": err3.Error(), "status": "404"}
+		return outputError
 	}
 
 	//se elimina registro en modificacion_vinculacion
 
 	if err4 := SendJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/modificacion_vinculacion/"+strconv.Itoa(respuesta_modificacion_vinculacion[0].Id), "DELETE", &respuesta_delete, respuesta_modificacion_vinculacion[0]); err4 == nil {
-		respuesta_total = "OK"
 	} else {
-		respuesta_total = "error"
-	}
+		outputError = map[string]interface{}{"funcion": "/AnularAdicionDocente2", "err": err4.Error(), "status": "404"}
+		return outputError	}
 
 	//Se elimina vinculacion nueva
 	if err5 := SendJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/vinculacion_docente/"+strconv.Itoa(v.DocentesDesvincular[0].Id), "DELETE", &respuesta_delete_vin, v.DocentesDesvincular[0]); err5 == nil {
 		fmt.Println("respuesta_eliminar_vin_nueva", respuesta_delete_vin)
-		respuesta_total = "OK"
 	} else {
-		respuesta_total = "error"
+		outputError = map[string]interface{}{"funcion": "/AnularAdicionDocente1", "err": err5.Error(), "status": "404"}
+		return outputError
 	}
 
-	return respuesta_total
+	return
 }
 
-func ConsultarCategoria(v models.VinculacionDocente) (respuesta string) {
+func ConsultarCategoria(v models.VinculacionDocente) (respuesta string, outputError map[string]interface{}) {
 
 	categoria, _, err := Buscar_Categoria_Docente(strconv.Itoa(v.VigenciaCarga), strconv.Itoa(v.PeriodoCarga), v.IdPersona)
 	if err != nil {
-		beego.Error(err)
+		return respuesta, err
 	}
 
 	respuesta = "OK"
@@ -143,10 +142,10 @@ func ConsultarCategoria(v models.VinculacionDocente) (respuesta string) {
 		respuesta = "Sin categoría"
 	}
 
-	return respuesta
+	return respuesta, nil
 }
 
-func ValidarSaldoCDP(validacion models.Objeto_Desvinculacion) (respuesta string) {
+func ValidarSaldoCDP(validacion models.Objeto_Desvinculacion) (respuesta string, outputError map[string]interface{}) {
 
 	var respuestaApropiacion models.DatosApropiacion
 	var saldoDisponibilidad float64
@@ -159,6 +158,7 @@ func ValidarSaldoCDP(validacion models.Objeto_Desvinculacion) (respuesta string)
 		respuesta = "OK"
 	} else {
 		respuesta = "Error"
+		return respuesta, err
 	}
 
 	valorContrato := validacion.DocentesDesvincular[0].ValorContrato
@@ -168,29 +168,33 @@ func ValidarSaldoCDP(validacion models.Objeto_Desvinculacion) (respuesta string)
 		respuesta = "OK"
 	} else {
 		respuesta = "Error"
+		outputError = map[string]interface{}{"funcion": "/ValidarSaldoCDP", "err": err2.Error(), "status": "404"}
+		return respuesta, outputError
 	}
 
 	if saldoDisponibilidad-valorContrato < 0 {
 		respuesta = "Error CDP"
 	}
-	return respuesta
+	return respuesta, nil
 }
 
-func AdicionarHoras(v models.Objeto_Desvinculacion) (respuesta string) {
+func AdicionarHoras(v models.Objeto_Desvinculacion) (respuesta string, outputError map[string]interface{}) {
 	var respuesta_mod_vin models.ModificacionVinculacion
 	var vinculacion_nueva int
+	var err1 map[string]interface{}
 	var temp_vinculacion [1]models.VinculacionDocente
 
 	//CAMBIAR ESTADO DE VINCULACIÓN DOCENTE
 	for _, pos := range v.DocentesDesvincular {
 		pos.NumeroRp = 0
 		pos.VigenciaRp = 0
-		err := SendJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/vinculacion_docente/"+strconv.Itoa(pos.Id), "PUT", &respuesta, pos)
-		//TODO: unificar errores
-		if err != nil {
-			err = fmt.Errorf("error al cambiar estado en vinculación docente al adicionar horas %s", err)
+		if err := SendJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/vinculacion_docente/"+strconv.Itoa(pos.Id), "PUT", &respuesta, pos); err == nil{
+		}else{
 			beego.Error(err)
+			outputError = map[string]interface{}{"funcion": "/AdicionarHoras1", "err": err.Error(), "status": "404"}
+			return "Error", outputError	
 		}
+		//TODO: unificar errores
 		beego.Info("respuesta", respuesta)
 		beego.Info("fechaAD", v.DocentesDesvincular[0].FechaInicioNueva)
 		temp_vinculacion[0] = models.VinculacionDocente{
@@ -212,9 +216,10 @@ func AdicionarHoras(v models.Objeto_Desvinculacion) (respuesta string) {
 		}
 
 		//CREAR NUEVA Vinculacion
-		vinculacion_nueva, err = InsertarDesvinculaciones(temp_vinculacion)
-		if err != nil {
-			beego.Error("error al realizar vinculacion nueva", err)
+		vinculacion_nueva, err1 = InsertarDesvinculaciones(temp_vinculacion)
+		if err1 != nil {
+			beego.Error("error al realizar vinculacion nueva", err1)
+			return "Error", err1	
 		}
 
 		//INSERCION  TABLA  DE TRAZA MODIFICACION VINCULACION
@@ -225,25 +230,26 @@ func AdicionarHoras(v models.Objeto_Desvinculacion) (respuesta string) {
 				VinculacionDocenteRegistrada: &models.VinculacionDocente{Id: vinculacion_nueva},
 				Horas:                        pos.NumeroHorasNuevas,
 			}
-			err := SendJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/modificacion_vinculacion/", "POST", &respuesta_mod_vin, temp)
-
-			if err != nil {
-				beego.Error("error en actualizacion de modificacion vinculacion de modificacion vinculacion", err)
-				respuesta = "error"
-			} else {
+			if err := SendJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/modificacion_vinculacion/", "POST", &respuesta_mod_vin, temp); err == nil {
 				beego.Info("respuesta modificacion vin", respuesta_mod_vin)
 				respuesta = "OK"
+			}else{
+				beego.Error("error en actualizacion de modificacion vinculacion de modificacion vinculacion", err)
+				outputError = map[string]interface{}{"funcion": "/AdicionarHoras2", "err": err.Error(), "status": "404"}
+				return "Error", outputError	
 			}
+
 		}
 	}
 
-	return respuesta
+	return respuesta, nil
 }
 
-func ActualizarVinculacionesCancelacion(v models.Objeto_Desvinculacion) (respuesta string) {
+func ActualizarVinculacionesCancelacion(v models.Objeto_Desvinculacion) (respuesta string, outputError map[string]interface{}) {
 	var respuesta_mod_vin models.ModificacionVinculacion
 	// var respuesta string
 	var vinculacion_nueva int
+	var err1 map[string]interface{}
 	var temp_vinculacion [1]models.VinculacionDocente
 
 	beego.Debug("para poner en false", v)
@@ -255,11 +261,12 @@ func ActualizarVinculacionesCancelacion(v models.Objeto_Desvinculacion) (respues
 
 		pos.NumeroRp = 0
 		pos.VigenciaRp = 0
-		err := SendJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/vinculacion_docente/"+strconv.Itoa(pos.Id), "PUT", &respuesta, pos)
-		if err != nil {
+		if err := SendJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/vinculacion_docente/"+strconv.Itoa(pos.Id), "PUT", &respuesta, pos); err == nil{
+		}else{
 			beego.Error("error en json", err)
+			outputError = map[string]interface{}{"funcion": "/ActualizarVinculacionesCancelacion1", "err": err.Error(), "status": "404"}
+			return "Error", outputError
 		}
-
 		//Verificar objeto para crear nuevas resoluciones
 		temp_vinculacion[0] = models.VinculacionDocente{
 			IdPersona:            pos.IdPersona,
@@ -279,9 +286,10 @@ func ActualizarVinculacionesCancelacion(v models.Objeto_Desvinculacion) (respues
 		}
 		fmt.Println("RP: ", temp_vinculacion[0].NumeroRp)
 		//CREAR NUEVA Vinculacion
-		vinculacion_nueva, err = InsertarDesvinculaciones(temp_vinculacion)
-		if err != nil {
-			beego.Error("error al realizar vinculacion nueva", err)
+		vinculacion_nueva, err1 = InsertarDesvinculaciones(temp_vinculacion)
+		if err1 != nil {
+			beego.Error("error al realizar vinculacion nueva", err1)
+			return "Error", err1	
 		}
 
 		//INSERCION  TABLA  DE TRAZA MODIFICACION VINCULACION
@@ -291,45 +299,45 @@ func ActualizarVinculacionesCancelacion(v models.Objeto_Desvinculacion) (respues
 			VinculacionDocenteRegistrada: &models.VinculacionDocente{Id: vinculacion_nueva},
 			Horas:                        pos.NumeroHorasSemanales,
 		}
-		errorMod := SendJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/modificacion_vinculacion/", "POST", &respuesta_mod_vin, temp)
-
-		if errorMod != nil {
-			beego.Error("error en actualizacion de modificacion vinculacion de modificacion vinculacion", err)
-			respuesta = "error"
-		} else {
+		if errorMod := SendJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/modificacion_vinculacion/", "POST", &respuesta_mod_vin, temp); errorMod == nil{
 			beego.Info("respuesta modificacion vin", respuesta_mod_vin)
 			respuesta = "OK"
+		}else{
+			beego.Error("error en actualizacion de modificacion vinculacion de modificacion vinculacion", errorMod)
+			outputError = map[string]interface{}{"funcion": "/ActualizarVinculacionesCancelacion2", "err": errorMod.Error(), "status": "404"}
+			return "Error", outputError	
 		}
 	}
 
-	return respuesta
+	return respuesta, nil
 
 }
 
-func InsertarDesvinculaciones(v [1]models.VinculacionDocente) (id int, err error) {
+func InsertarDesvinculaciones(v [1]models.VinculacionDocente) (id int, outputError map[string]interface{}) {
 	var d []models.VinculacionDocente
-	json_ejemplo, err := json.Marshal(v)
-	if err != nil {
-		beego.Error(err)
-		return id, err
+	json_ejemplo, err1 := json.Marshal(v)
+	if err1 != nil {
+		beego.Error(err1)
+		outputError = map[string]interface{}{"funcion": "/InsertarDesvinculaciones", "err": "Error en codificación de datos", "status": "404"}
+		return id, outputError
 	}
-	err = json.Unmarshal(json_ejemplo, &d)
+	err2 := json.Unmarshal(json_ejemplo, &d)
 
-	if err != nil {
-		beego.Error(err)
-		return id, err
+	if err2 != nil {
+		beego.Error(err2)
+		outputError = map[string]interface{}{"funcion": "/InsertarDesvinculaciones", "err": "Error en decodificación de datos", "status": "404"}
+		return id, outputError
 	}
 
 	//TODO: unificar cont con error
-	d, err = CalcularSalarioPrecontratacion(d)
+	d, err := CalcularSalarioPrecontratacion(d)
 	if err != nil {
 		return id, err
 	}
 
-	err = SendJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/vinculacion_docente/InsertarVinculaciones/", "POST", &id, &d)
-	if err != nil {
-		beego.Error(err)
-		return 0, err
+	if err := SendJson(beego.AppConfig.String("ProtocolAdmin")+"://"+beego.AppConfig.String("UrlcrudAdmin")+"/"+beego.AppConfig.String("NscrudAdmin")+"/vinculacion_docente/InsertarVinculaciones/", "POST", &id, &d); err == nil{
+		outputError = map[string]interface{}{"funcion": "/InsertarDesvinculaciones", "err": err.Error(), "status": "404"}
+		return 0, outputError
 	}
-	return id, err
+	return id, nil
 }
